Updated page.tsx
