+++ Frontend/Gym-Frontend/gym-trainer/app/doctor/video-session/page.tsx
@@ -9,7 +9,15 @@ import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Input } from "@/components/ui/input";
-import { Mic, MicOff, VideoIcon, VideoOff, Phone, Send } from "lucide-react";
+import {
+  Mic,
+  MicOff,
+  VideoIcon,
+  VideoOff,
+  Phone,
+  Send,
+  Stethoscope,
+} from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
 
 interface Message {
@@ -20,7 +28,7 @@ interface Message {
   isUser: boolean;
 }
 
-export default function VideoSessionPage() {
+export default function DoctorVideoConsultationPage() {
   const searchParams = useSearchParams();
   const sessionId = searchParams.get("session");
   const { toast } = useToast();
@@ -32,7 +40,7 @@ export default function VideoSessionPage() {
   const [messages, setMessages] = useState<Message[]>([
     {
       id: "1",
-      sender: "John Smith",
+      sender: "Dr. Smith",
       content: "Hello! How are you feeling today?",
       timestamp: new Date(new Date().setHours(10, 1)),
       isUser: false,
@@ -40,14 +48,15 @@ export default function VideoSessionPage() {
     {
       id: "2",
       sender: "You",
-      content: "Hi John! I'm good, ready for our session.",
+      content:
+        "Hi Dr. Smith! I've been having some symptoms I'd like to discuss.",
       timestamp: new Date(new Date().setHours(10, 2)),
       isUser: true,
     },
     {
       id: "3",
-      sender: "John Smith",
-      content: "Great! Let's start with some warm-up exercises.",
+      sender: "Dr. Smith",
+      content: "I understand. Let's go through your symptoms one by one.",
       timestamp: new Date(new Date().setHours(10, 3)),
       isUser: false,
     },
@@ -66,7 +75,7 @@ export default function VideoSessionPage() {
       if (sessionId) {
         try {
           const response = await fetch(
-            `/api/video-session?sessionId=${sessionId}`
+            `/api/doctor-video-session?sessionId=${sessionId}`
           );
           if (response.ok) {
             const data = await response.json();
@@ -158,7 +167,7 @@ export default function VideoSessionPage() {
       const tracks = stream.getTracks();
       tracks.forEach((track) => track.stop());
     }
-    window.location.href = "/gym-trainer";
+    window.location.href = "/doctor";
   };
 
   const sendMessage = (e: React.FormEvent) => {
@@ -174,16 +183,17 @@ export default function VideoSessionPage() {
       setMessages([...messages, newMessage]);
       setMessage("");
 
-      // Simulate trainer response
+      // Simulate doctor response
       setTimeout(() => {
-        const trainerResponse: Message = {
+        const doctorResponse: Message = {
           id: (Date.now() + 1).toString(),
-          sender: "John Smith",
-          content: "I see. Let's continue with the next exercise.",
+          sender: "Dr. Smith",
+          content:
+            "Thank you for sharing that information. Based on your symptoms, I recommend...",
           timestamp: new Date(),
           isUser: false,
         };
-        setMessages((prev) => [...prev, trainerResponse]);
+        setMessages((prev) => [...prev, doctorResponse]);
       }, 2000);
     }
   };
@@ -192,15 +202,15 @@ export default function VideoSessionPage() {
     return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
   };
 
-  const trainerName = sessionData?.appointment?.trainer || "John Smith";
+  const doctorName = sessionData?.appointment?.doctor || "Dr. Smith";
 
   return (
     <div className="container mx-auto p-4 max-w-6xl">
       <div className="flex items-center mb-8">
-        <Link href="/gym-trainer" className="mr-4">
+        <Link href="/doctor" className="mr-4">
           <Button variant="outline">Back to Appointments</Button>
         </Link>
-        <h1 className="text-3xl font-bold">Video Session</h1>
+        <h1 className="text-3xl font-bold">Medical Consultation</h1>
       </div>
 
       <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
@@ -225,9 +235,9 @@ export default function VideoSessionPage() {
                 </div>
               )}
               <div className="absolute top-4 left-4 bg-black/70 text-white px-3 py-1 rounded-full text-sm">
-                Live Session with {trainerName}
+                Medical Consultation with {doctorName}
               </div>
-              <div className="absolute bottom-4 right-4 w-32 h-24 bg-gray-800 rounded-md overflow-hidden border-2 border-primary">
+              <div className="absolute bottom-4 right-4 w-32 h-24 bg-gray-800 rounded-md overflow-hidden border-2 border-sky-500">
                 <video
                   autoPlay
                   muted
@@ -241,7 +251,7 @@ export default function VideoSessionPage() {
               <Button
                 variant={isMicOn ? "default" : "destructive"}
                 size="icon"
-                className="rounded-full h-12 w-12"
+                className="rounded-full h-12 w-12 bg-sky-500 hover:bg-sky-600"
                 onClick={toggleMic}
               >
                 {isMicOn ? <Mic /> : <MicOff />}
@@ -249,7 +259,7 @@ export default function VideoSessionPage() {
               <Button
                 variant={isVideoOn ? "default" : "destructive"}
                 size="icon"
-                className="rounded-full h-12 w-12"
+                className="rounded-full h-12 w-12 bg-sky-500 hover:bg-sky-600"
                 onClick={toggleVideo}
               >
                 {isVideoOn ? <VideoIcon /> : <VideoOff />}
@@ -288,7 +298,7 @@ export default function VideoSessionPage() {
                       <div
                         className={`max-w-[80%] rounded-2xl p-3 ${
                           msg.isUser
-                            ? "bg-primary text-primary-foreground"
+                            ? "bg-sky-500 text-white"
                             : "bg-muted text-foreground"
                         }`}
                       >
@@ -314,7 +324,11 @@ export default function VideoSessionPage() {
                       placeholder="Type a message..."
                       className="flex-1"
                     />
-                    <Button type="submit" size="icon" className="bg-sky-500">
+                    <Button
+                      type="submit"
+                      size="icon"
+                      className="bg-sky-500 hover:bg-sky-600"
+                    >
                       <Send className="h-4 w-4" />
                     </Button>
                   </div>
@@ -326,23 +340,23 @@ export default function VideoSessionPage() {
               >
                 <div className="space-y-4">
                   <div className="flex items-center space-x-3 p-2 rounded-md bg-muted">
-                    <div className="w-10 h-10 rounded-full bg-500 flex items-center justify-center">
-                      <VideoIcon className="h-5 w-5 text-primary" />
+                    <div className="w-10 h-10 rounded-full bg-sky-100 flex items-center justify-center">
+                      <Stethoscope className="h-5 w-5 text-sky-500" />
                     </div>
                     <div>
-                      <p className="font-medium">{trainerName}</p>
+                      <p className="font-medium">{doctorName}</p>
                       <p className="text-sm text-muted-foreground">
-                        Trainer (Host)
+                        Doctor (Host)
                       </p>
                     </div>
                   </div>
                   <div className="flex items-center space-x-3 p-2 rounded-md bg-muted">
-                    <div className="w-10 h-10 rounded-full bg-primary/20 flex items-center justify-center">
-                      <VideoIcon className="h-5 w-5 text-primary" />
+                    <div className="w-10 h-10 rounded-full bg-sky-100 flex items-center justify-center">
+                      <VideoIcon className="h-5 w-5 text-sky-500" />
                     </div>
                     <div>
                       <p className="font-medium">You</p>
-                      <p className="text-sm text-muted-foreground">Client</p>
+                      <p className="text-sm text-muted-foreground">Patient</p>
                     </div>
                   </div>
                 </div>
