+++ Frontend/Gym-Frontend/gym-trainer/app/gym-trainer/video-session/loading.tsx
@@ -1,4 +1,3 @@
 export default function Loading() {
-  return null
+  return null;
 }
-
