+++ Frontend/Gym-Frontend/gym-trainer/app/gym-trainer/video-session/page.tsx
@@ -9,15 +9,7 @@ import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Input } from "@/components/ui/input";
-import {
-  Mic,
-  MicOff,
-  VideoIcon,
-  VideoOff,
-  Phone,
-  Send,
-  Stethoscope,
-} from "lucide-react";
+import { Mic, MicOff, VideoIcon, VideoOff, Phone, Send } from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
 
 interface Message {
@@ -28,7 +20,7 @@ interface Message {
   isUser: boolean;
 }
 
-export default function DoctorVideoConsultationPage() {
+export default function VideoSessionPage() {
   const searchParams = useSearchParams();
   const sessionId = searchParams.get("session");
   const { toast } = useToast();
@@ -40,7 +32,7 @@ export default function DoctorVideoConsultationPage() {
   const [messages, setMessages] = useState<Message[]>([
     {
       id: "1",
-      sender: "Dr. Smith",
+      sender: "John Smith",
       content: "Hello! How are you feeling today?",
       timestamp: new Date(new Date().setHours(10, 1)),
       isUser: false,
@@ -48,15 +40,14 @@ export default function DoctorVideoConsultationPage() {
     {
       id: "2",
       sender: "You",
-      content:
-        "Hi Dr. Smith! I've been having some symptoms I'd like to discuss.",
+      content: "Hi John! I'm good, ready for our session.",
       timestamp: new Date(new Date().setHours(10, 2)),
       isUser: true,
     },
     {
       id: "3",
-      sender: "Dr. Smith",
-      content: "I understand. Let's go through your symptoms one by one.",
+      sender: "John Smith",
+      content: "Great! Let's start with some warm-up exercises.",
       timestamp: new Date(new Date().setHours(10, 3)),
       isUser: false,
     },
@@ -75,7 +66,7 @@ export default function DoctorVideoConsultationPage() {
       if (sessionId) {
         try {
           const response = await fetch(
-            `/api/doctor-video-session?sessionId=${sessionId}`
+            `/api/video-session?sessionId=${sessionId}`
           );
           if (response.ok) {
             const data = await response.json();
@@ -167,7 +158,7 @@ export default function DoctorVideoConsultationPage() {
       const tracks = stream.getTracks();
       tracks.forEach((track) => track.stop());
     }
-    window.location.href = "/doctor";
+    window.location.href = "/gym-trainer";
   };
 
   const sendMessage = (e: React.FormEvent) => {
@@ -183,17 +174,16 @@ export default function DoctorVideoConsultationPage() {
       setMessages([...messages, newMessage]);
       setMessage("");
 
-      // Simulate doctor response
+      // Simulate trainer response
       setTimeout(() => {
-        const doctorResponse: Message = {
+        const trainerResponse: Message = {
           id: (Date.now() + 1).toString(),
-          sender: "Dr. Smith",
-          content:
-            "Thank you for sharing that information. Based on your symptoms, I recommend...",
+          sender: "John Smith",
+          content: "I see. Let's continue with the next exercise.",
           timestamp: new Date(),
           isUser: false,
         };
-        setMessages((prev) => [...prev, doctorResponse]);
+        setMessages((prev) => [...prev, trainerResponse]);
       }, 2000);
     }
   };
@@ -202,15 +192,15 @@ export default function DoctorVideoConsultationPage() {
     return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
   };
 
-  const doctorName = sessionData?.appointment?.doctor || "Dr. Smith";
+  const trainerName = sessionData?.appointment?.trainer || "John Smith";
 
   return (
     <div className="container mx-auto p-4 max-w-6xl">
       <div className="flex items-center mb-8">
-        <Link href="/doctor" className="mr-4">
+        <Link href="/gym-trainer" className="mr-4">
           <Button variant="outline">Back to Appointments</Button>
         </Link>
-        <h1 className="text-3xl font-bold">Medical Consultation</h1>
+        <h1 className="text-3xl font-bold">Video Session</h1>
       </div>
 
       <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
@@ -235,9 +225,9 @@ export default function DoctorVideoConsultationPage() {
                 </div>
               )}
               <div className="absolute top-4 left-4 bg-black/70 text-white px-3 py-1 rounded-full text-sm">
-                Medical Consultation with {doctorName}
+                Live Session with {trainerName}
               </div>
-              <div className="absolute bottom-4 right-4 w-32 h-24 bg-gray-800 rounded-md overflow-hidden border-2 border-sky-500">
+              <div className="absolute bottom-4 right-4 w-32 h-24 bg-gray-800 rounded-md overflow-hidden border-2 border-primary">
                 <video
                   autoPlay
                   muted
@@ -251,7 +241,7 @@ export default function DoctorVideoConsultationPage() {
               <Button
                 variant={isMicOn ? "default" : "destructive"}
                 size="icon"
-                className="rounded-full h-12 w-12 bg-sky-500 hover:bg-sky-600"
+                className="rounded-full h-12 w-12"
                 onClick={toggleMic}
               >
                 {isMicOn ? <Mic /> : <MicOff />}
@@ -259,7 +249,7 @@ export default function DoctorVideoConsultationPage() {
               <Button
                 variant={isVideoOn ? "default" : "destructive"}
                 size="icon"
-                className="rounded-full h-12 w-12 bg-sky-500 hover:bg-sky-600"
+                className="rounded-full h-12 w-12"
                 onClick={toggleVideo}
               >
                 {isVideoOn ? <VideoIcon /> : <VideoOff />}
@@ -298,7 +288,7 @@ export default function DoctorVideoConsultationPage() {
                       <div
                         className={`max-w-[80%] rounded-2xl p-3 ${
                           msg.isUser
-                            ? "bg-sky-500 text-white"
+                            ? "bg-primary text-primary-foreground"
                             : "bg-muted text-foreground"
                         }`}
                       >
@@ -324,11 +314,7 @@ export default function DoctorVideoConsultationPage() {
                       placeholder="Type a message..."
                       className="flex-1"
                     />
-                    <Button
-                      type="submit"
-                      size="icon"
-                      className="bg-sky-500 hover:bg-sky-600"
-                    >
+                    <Button type="submit" size="icon" className="bg-sky-500">
                       <Send className="h-4 w-4" />
                     </Button>
                   </div>
@@ -340,23 +326,23 @@ export default function DoctorVideoConsultationPage() {
               >
                 <div className="space-y-4">
                   <div className="flex items-center space-x-3 p-2 rounded-md bg-muted">
-                    <div className="w-10 h-10 rounded-full bg-sky-100 flex items-center justify-center">
-                      <Stethoscope className="h-5 w-5 text-sky-500" />
+                    <div className="w-10 h-10 rounded-full bg-500 flex items-center justify-center">
+                      <VideoIcon className="h-5 w-5 text-primary" />
                     </div>
                     <div>
-                      <p className="font-medium">{doctorName}</p>
+                      <p className="font-medium">{trainerName}</p>
                       <p className="text-sm text-muted-foreground">
-                        Doctor (Host)
+                        Trainer (Host)
                       </p>
                     </div>
                   </div>
                   <div className="flex items-center space-x-3 p-2 rounded-md bg-muted">
-                    <div className="w-10 h-10 rounded-full bg-sky-100 flex items-center justify-center">
-                      <VideoIcon className="h-5 w-5 text-sky-500" />
+                    <div className="w-10 h-10 rounded-full bg-primary/20 flex items-center justify-center">
+                      <VideoIcon className="h-5 w-5 text-primary" />
                     </div>
                     <div>
                       <p className="font-medium">You</p>
-                      <p className="text-sm text-muted-foreground">Patient</p>
+                      <p className="text-sm text-muted-foreground">Client</p>
                     </div>
                   </div>
                 </div>
